Shifted several `IO` companion functions to share with Native
